move shopping item from cookies to DB after login

+++ HHMarketWebApp/HHMarketWebApp/Models/CartItem.cs
@@ -9,6 +9,6 @@ namespace HHMarketWebApp.Models
     {
         public decimal Price { get; set; }
         public int ProductDetailsId { get; set; } 
-        public int Amount { get; set; }
+        public short Amount { get; set; }
     }
 }
